# openapi.yml
{"openapi":"3.0.2","info":{"title":"API Documentation","version":"1.0.0"},"paths":{"/jokes":{"get":{"summary":"Jokes","operationId":"jokes_jokes_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Insert Joke","operationId":"insert_joke_jokes_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JokeCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JokeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Joke","operationId":"delete_joke_jokes_delete","parameters":[{"required":true,"schema":{"title":"Number","type":"integer"},"name":"number","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"summary":"Update Joke","operationId":"update_joke_jokes_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JokeUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JokeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jokes/{source}":{"get":{"summary":"Jokes Source","operationId":"jokes_source_jokes__source__get","parameters":[{"required":true,"schema":{"title":"Source","type":"string"},"name":"source","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/least-common-multiple":{"get":{"summary":"Least Common Multiple","operationId":"least_common_multiple_least_common_multiple_get","parameters":[{"required":true,"schema":{"title":"Numbers","type":"array","items":{"type":"integer"}},"example":"/least-common-multiple?numbers=4&numbersddd=6&numbesssrs=8","name":"numbers","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/addition":{"get":{"summary":"Addition","operationId":"addition_addition_get","parameters":[{"required":true,"schema":{"title":"Number","type":"integer"},"example":"/addition?number=5","name":"number","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"JokeCreate":{"title":"JokeCreate","required":["joke"],"type":"object","properties":{"joke":{"title":"Joke","maxLength":500,"minLength":1,"type":"string"}}},"JokeResponse":{"title":"JokeResponse","required":["joke","number"],"type":"object","properties":{"joke":{"title":"Joke","maxLength":500,"minLength":1,"type":"string"},"number":{"title":"Number","type":"integer"}}},"JokeUpdate":{"title":"JokeUpdate","required":["joke","number"],"type":"object","properties":{"joke":{"title":"Joke","maxLength":500,"minLength":1,"type":"string"},"number":{"title":"Number","type":"integer"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}